+++ packages/vsce/src/commands/closeLocalFileCommand.ts
@@ -12,109 +12,88 @@
 import { CicsCmciConstants, ICMCIApiResponse } from "@zowe/cics-for-zowe-sdk";
 import { imperative } from "@zowe/zowe-explorer-api";
 import { ProgressLocation, TreeView, commands, window } from "vscode";
-import { CICSCombinedLocalFileTree } from "../trees/CICSCombinedTrees/CICSCombinedLocalFileTree";
-import { CICSRegionTree } from "../trees/CICSRegionTree";
-import { CICSRegionsContainer } from "../trees/CICSRegionsContainer";
-import { CICSTree } from "../trees/CICSTree";
-import { CICSLocalFileTreeItem } from "../trees/treeItems/CICSLocalFileTreeItem";
-import { findSelectedNodes, splitCmciErrorMessage } from "../utils/commandUtils";
 import constants from "../constants/CICS.defaults";
+import { ILocalFile, LocalFileMeta } from "../doc";
+import { CICSResourceContainerNode } from "../trees";
+import { CICSTree } from "../trees/CICSTree";
+import { findSelectedNodes, splitCmciErrorMessage } from "../utils/commandUtils";
 import { runPutResource } from "../utils/resourceUtils";
 import { ICommandParams } from "./ICommandParams";
 
 export function getCloseLocalFileCommand(tree: CICSTree, treeview: TreeView<any>) {
   return commands.registerCommand("cics-extension-for-zowe.closeLocalFile", async (clickedNode) => {
-    const allSelectedNodes = findSelectedNodes(treeview, CICSLocalFileTreeItem, clickedNode);
-    if (!allSelectedNodes || !allSelectedNodes.length) {
+    const nodes = findSelectedNodes(treeview, LocalFileMeta, clickedNode);
+    if (!nodes || !nodes.length) {
       window.showErrorMessage("No CICS local file selected");
       return;
     }
-    const parentRegions: CICSRegionTree[] = [];
+
     let busyDecision = await window.showInformationMessage(
       `Choose one of the following for the file busy condition`,
       ...["Wait", "No Wait", "Force"]
     );
-    if (busyDecision) {
-      busyDecision = busyDecision.replace(" ", "").toUpperCase();
+    if (!busyDecision) {
+      return;
+    }
+    busyDecision = busyDecision.replace(" ", "").toUpperCase();
 
-      window.withProgress(
-        {
-          title: "Close",
-          location: ProgressLocation.Notification,
-          cancellable: true,
-        },
-        async (progress, token) => {
-          token.onCancellationRequested(() => {});
-          for (const index in allSelectedNodes) {
-            progress.report({
-              message: `Closing ${parseInt(index) + 1} of ${allSelectedNodes.length}`,
-              increment: (parseInt(index) / allSelectedNodes.length) * constants.PERCENTAGE_MAX,
-            });
-            const currentNode = allSelectedNodes[parseInt(index)];
+    window.withProgress(
+      {
+        title: "Close",
+        location: ProgressLocation.Notification,
+        cancellable: true,
+      },
+      async (progress, token) => {
+        token.onCancellationRequested(() => {});
 
-            try {
-              await closeLocalFile(
-                currentNode.parentRegion.parentSession.session,
-                {
-                  name: currentNode.localFile.file,
-                  regionName: currentNode.parentRegion.label,
-                  cicsPlex: currentNode.parentRegion.parentPlex ? currentNode.parentRegion.parentPlex.getPlexName() : undefined,
-                },
-                busyDecision
-              );
-              if (!parentRegions.includes(currentNode.parentRegion)) {
-                parentRegions.push(currentNode.parentRegion);
-              }
-            } catch (error) {
+        for (const node of nodes) {
+          progress.report({
+            message: `Closing ${nodes.indexOf(node) + 1} of ${nodes.length}`,
+            increment: (nodes.indexOf(node) / nodes.length) * constants.PERCENTAGE_MAX,
+          });
+
+          try {
+            await closeLocalFile(
+              node.getSession(),
+              {
+                name: node.getContainedResourceName(),
+                regionName: node.regionName,
+                cicsPlex: node.cicsplexName,
+              },
+              busyDecision
+            );
+          } catch (error) {
+            // @ts-ignore
+            if (error.mMessage) {
               // @ts-ignore
-              if (error.mMessage) {
-                // @ts-ignore
-                const [_resp, resp2, respAlt, eibfnAlt] = splitCmciErrorMessage(error.mMessage);
+              const [_resp, resp2, respAlt, eibfnAlt] = splitCmciErrorMessage(error.mMessage);
 
-                window.showErrorMessage(
-                  `Perform CLOSE on local file "${
-                    allSelectedNodes[parseInt(index)].localFile.file
-                  }" failed: EXEC CICS command (${eibfnAlt}) RESP(${respAlt}) RESP2(${resp2})`
-                );
-              } else {
-                window.showErrorMessage(
-                  `Something went wrong when performing a CLOSE - ${JSON.stringify(error, Object.getOwnPropertyNames(error)).replace(
-                    /(\\n\t|\\n|\\t)/gm,
-                    " "
-                  )}`
-                );
-              }
-            }
-          }
-          for (const parentRegion of parentRegions) {
-            try {
-              const localFileTree = parentRegion.children.filter((child: any) => child.contextValue.includes("cicstreelocalfile."))[0];
-              // Only load contents if the tree is expanded
-              if (localFileTree.collapsibleState === 2) {
-                await localFileTree.loadContents();
-              }
-              // if node is in a plex and the plex contains the region container tree
-              if (parentRegion.parentPlex && parentRegion.parentPlex.children.some((child) => child instanceof CICSRegionsContainer)) {
-                const allLocalFileTreeTree = parentRegion.parentPlex.children.filter((child: any) =>
-                  child.contextValue.includes("cicscombinedlocalfiletree.")
-                )[0] as CICSCombinedLocalFileTree;
-                if (allLocalFileTreeTree.collapsibleState === 2 && allLocalFileTreeTree.getActiveFilter()) {
-                  await allLocalFileTreeTree.loadContents(tree);
-                }
-              }
-            } catch (error) {
               window.showErrorMessage(
-                `Something went wrong when reloading local files - ${JSON.stringify(error, Object.getOwnPropertyNames(error)).replace(
+                `Perform CLOSE on local file "${node
+                  .getContainedResource()
+                  .meta.getName(node.getContainedResource().resource)}" failed: EXEC CICS command (${eibfnAlt}) RESP(${respAlt}) RESP2(${resp2})`
+              );
+            } else {
+              window.showErrorMessage(
+                `Something went wrong when performing a CLOSE - ${JSON.stringify(error, Object.getOwnPropertyNames(error)).replace(
                   /(\\n\t|\\n|\\t)/gm,
                   " "
                 )}`
               );
             }
           }
-          tree._onDidChangeTreeData.fire(undefined);
         }
-      );
-    }
+        // Work out how many files to re-fetch
+        const parentNode = nodes[0].getParent() as CICSResourceContainerNode<ILocalFile>;
+        let numToFetch = parentNode.children.length;
+        if (!parentNode.getChildResource().resources.getFetchedAll()) {
+          numToFetch -= 1;
+        }
+        parentNode.getChildResource().resources.setNumberToFetch(numToFetch);
+
+        tree._onDidChangeTreeData.fire(parentNode);
+      }
+    );
   });
 }