+++ packages/vsce/src/webviews/resource-inspector-panel/index.tsx
@@ -11,8 +11,8 @@
 
 import * as React from "react";
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import ResourceInspector from './ResourceInspector';
 
 const container = document.getElementById('webviewRoot');
 const root = createRoot(container);
-root.render(<App />);
+root.render(<ResourceInspector />);