+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the IBM® CICS® Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+
+- Add missing CICS resource constants. [#249](https://github.com/zowe/cics-for-zowe-client/pull/249)
+
 ## `6.8.0`
 
 - Enhancement: Added extender API interface and resource constants for pipeline and webservice resource. [#299](https://github.com/zowe/cics-for-zowe-client/pull/299)